src/requirements-dev.txt
pytest>=8.0
playwright
httpx
Pillow
